feat: :sparkles: reuse checkForCityParams error check, change naming for future third route
