+++ apps/backend/package.json
@@ -20,10 +20,11 @@
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
     "express-async-handler": "^1.2.0",
+    "helmet": "^7.1.0",
+    "morgan": "^1.10.0",
     "shared-schemas": "workspace:*",
     "shared-types": "workspace:*",
-    "helmet": "^7.1.0",
-    "morgan": "^1.10.0"
+    "zod": "^3.23.8"
   },
   "devDependencies": {
     "@types/cors": "^2.8.17",
