+++ apps/backend/src/controllers/currentWeatherController.ts
@@ -1,9 +1,9 @@
-import { NextFunction, Request, Response } from 'express';
+import { Request, Response } from 'express';
 import axios from 'axios';
 import { FetchTodayOverviewResult } from 'shared-schemas/apiSchemas';
 import { prisma } from '../prismaClient';
 
-const flattenOverviewObject = (parsedData: any) => {
+const flattenCurrentWeatherObject = (parsedData: any) => {
   const finalObj: any = {};
 
   const flatten = (obj: any) => {
@@ -20,16 +20,7 @@ const flattenOverviewObject = (parsedData: any) => {
   return finalObj;
 };
 
-export const checkForQueryErrors = (
-  req: Request,
-  _res: Response,
-  next: NextFunction
-) => {
-  const { city } = req.query;
-  !city ? next(new Error('City query parameter is required!')) : next();
-};
-
-export const getTodayOverview = async (req: Request, res: Response) => {
+export const getCurrentWeather = async (req: Request, res: Response) => {
   const wData = await prisma.weatherData.findFirst();
 
   const { city } = req.query;
@@ -42,8 +33,8 @@ export const getTodayOverview = async (req: Request, res: Response) => {
   });
 
   const parsedData = FetchTodayOverviewResult.parse(data);
-  const flattenedData = flattenOverviewObject(parsedData);
+  const flattenedData = flattenCurrentWeatherObject(parsedData);
   // await prisma.weatherData.create({ data: flattenedData });
 
-  res.json(flattenOverviewObject(wData));
+  res.json(flattenCurrentWeatherObject(wData));
 };
