+++ apps/backend/src/middlewares/errorCheckers.ts
@@ -1,0 +1,34 @@
+import { NextFunction, Request, Response } from 'express';
+
+import ErrorResponse from '../interfaces/ErrorResponse';
+
+export const checkForCityParameter = (
+  req: Request,
+  _res: Response,
+  next: NextFunction
+) => {
+  const { city } = req.query;
+  !city ? next(new Error('City query parameter is required!')) : next();
+};
+
+export function notFound(req: Request, res: Response, next: NextFunction) {
+  res.status(404);
+  const error = new Error(`🔍 - Not Found - ${req.originalUrl}`);
+  next(error);
+}
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export function errorHandler(
+  err: Error,
+  _req: Request,
+  res: Response<ErrorResponse>,
+  _next: NextFunction
+) {
+  const statusCode = res.statusCode !== 200 ? res.statusCode : 500;
+  res.status(statusCode);
+  res.json({
+    status: statusCode,
+    message: err.message,
+    stack: process.env.NODE_ENV === 'production' ? '🥞' : err.stack,
+  });
+}
