+++ apps/backend/src/routes/currentWeatherRoutes.ts
@@ -1,0 +1,8 @@
+import { Router } from 'express';
+import { getCurrentWeather } from '../controllers/currentWeatherController';
+
+const currentWeatherRouter = Router();
+
+currentWeatherRouter.route('/').get(getCurrentWeather);
+
+export default currentWeatherRouter;
